pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-limiter>=3.10.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-restful>=0.3.10",
    "flask-swagger-ui>=4.11.1",
    "flask-login>=0.6.3",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.38",
]
